FEAT: creates custom hooks and fixes repeated token error logic in editToken form
