+++ src/components/AddTokenForm.jsx
@@ -1,66 +1,20 @@
-import React, { useState, useEffect, useContext } from 'react';
-import { useNavigate } from 'react-router-dom';
+import React from 'react';
 import { Button } from 'react-bootstrap';
-import AppContext from '../context/AppContext';
+import { useAddTokenForm } from '../hooks';
 import WishWalletTitle from './WishWalletTitle';
 import 'bootstrap/dist/css/bootstrap.min.css';
 
 const AddTokenForm = () => {
-  const { tokens, setTokens } = useContext(AppContext);
-  const [token, setToken] = useState('');
-  const [balance, setBalance] = useState('');
-  const [repeatedToken, setRepeatedToken] = useState(false);
-  const [emptyToken, setEmptyToken] = useState(false);
-  const [emptyBalance, setEmptyBalance] = useState(false);
-  const [entriesAreValid, setEntriesAreValid] = useState(false);
-  const navigate = useNavigate();
-  const tokensStorage = JSON.parse(localStorage.getItem('tokens'));
-  const tokenList = tokensStorage || tokens;
-
-  useEffect(() => {
-    const tokenIsRepeated = tokenList.some((obj) => obj.token === token);
-
-    if (tokenIsRepeated || token === '' || balance === '') {
-      setEntriesAreValid(false);
-    } else {
-      setEntriesAreValid(true);
-    }
-  }, [token, balance, tokenList]);
-
-  const handleChange = ({ target: { name, value } }) => {
-    if (name === 'token-input') setToken(value);
-    if (name === 'balance-input') setBalance(value);
-  };
-
-  const handleErrorMessages = () => {
-    const tokenIsRepeated = tokenList.some((obj) => obj.token === token);
-
-    if (tokenIsRepeated) {
-      setRepeatedToken(true);
-    } else {
-      setRepeatedToken(false);
-    }
-    if (token.length === 0) {
-      setEmptyToken(true);
-    } else {
-      setEmptyToken(false);
-    }
-    if (balance.length === 0) {
-      setEmptyBalance(true);
-    } else {
-      setEmptyBalance(false);
-    }
-  };
-
-  const handleSave = () => {
-    if (entriesAreValid) {
-      const newTokens = [...tokenList, { token, balance }];
-      localStorage.setItem('tokens', JSON.stringify(newTokens));
-      setTokens(newTokens);
-      return navigate('/');
-    }
-    return handleErrorMessages();
-  };
+  const {
+    token,
+    balance,
+    repeatedToken,
+    emptyToken,
+    emptyBalance,
+    navigate,
+    handleChange,
+    handleSave,
+  } = useAddTokenForm();
 
   return (
     <main className="container">
